move custom tg as well and lint
